Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/mediacapture-streams/MediaStream-default-feature-policy.https-expected.txt

CONSOLE MESSAGE: line 13: Trying to call getUserMedia from a frame without correct 'allow' attribute.
CONSOLE MESSAGE: line 13: Trying to call getUserMedia from a frame without correct 'allow' attribute.
CONSOLE MESSAGE: line 13: Trying to call getUserMedia from a frame without correct 'allow' attribute.
CONSOLE MESSAGE: line 13: Trying to call getUserMedia from a frame without correct 'allow' attribute.
CONSOLE MESSAGE: line 13: Trying to call getUserMedia from a frame without correct 'allow' attribute.
CONSOLE MESSAGE: line 13: Trying to call getUserMedia from a frame without correct 'allow' attribute.
CONSOLE MESSAGE: line 13: Trying to call getUserMedia from a frame without correct 'allow' attribute.
CONSOLE MESSAGE: line 13: Trying to call getUserMedia from a frame without correct 'allow' attribute.
CONSOLE MESSAGE: line 13: Trying to call getUserMedia from a frame without correct 'allow' attribute.


PASS Default "microphone" feature policy ["self"] allows the top-level document. 
PASS Default "microphone" feature policy ["self"] allows same-origin iframes. 
PASS Default "microphone" feature policy ["self"] disallows cross-origin iframes. 
PASS Feature policy "microphone" can be enabled in cross-origin iframes using "allow" attribute. 
PASS Default "camera" feature policy ["self"] allows the top-level document. 
PASS Default "camera" feature policy ["self"] allows same-origin iframes. 
PASS Default "camera" feature policy ["self"] disallows cross-origin iframes. 
PASS Feature policy "camera" can be enabled in cross-origin iframes using "allow" attribute. 
PASS Default "camera;microphone" feature policy ["self"] allows the top-level document. 
PASS Default "camera;microphone" feature policy ["self"] allows same-origin iframes. 
PASS Default "camera;microphone" feature policy ["self"] disallows cross-origin iframes. 
FAIL Feature policy "camera;microphone" can be enabled in cross-origin iframes using "allow" attribute. assert_equals: expected "#OK" but got "#audio undefined and video true constraints must not be allowed."
